Move test for failing halfway through batch to DynamoDB-specific class
