+++ java/core/src/test/java/sleeper/core/statestore/inmemory/InMemoryFileReferenceStoreTest.java
@@ -503,24 +503,6 @@ public class InMemoryFileReferenceStoreTest extends InMemoryStateStoreTestBase {
         }
 
         @Test
-        public void shouldFailWhenSameFileIsAssignedInDifferentRequests() throws Exception {
-            // Given
-            FileReference file1 = factory.rootFile("file1", 100L);
-            FileReference file2 = factory.rootFile("file2", 100L);
-            store.addFiles(List.of(file1, file2));
-
-            // When / Then
-            assertThatThrownBy(() -> store.atomicallyAssignJobIdsToFileReferences(List.of(
-                    assignJobOnPartitionToFiles("job1", "root", List.of("file1")),
-                    assignJobOnPartitionToFiles("job2", "root", List.of("file1")))))
-                    .isInstanceOf(FileReferenceAssignedToJobException.class);
-            assertThat(store.getFileReferences()).containsExactly(
-                    withJobId("job1", file1),
-                    file2);
-            assertThat(store.getFileReferencesWithNoJobId()).containsExactly(file2);
-        }
-
-        @Test
         public void shouldNotMarkFileWithJobIdWhenOneIsAlreadySet() throws Exception {
             // Given
             FileReference file = factory.rootFile("file", 100L);
